Move exception package to the enforcer common module Fix code review suggestions
