+++ enforcer-parent/enforcer/src/main/java/org/wso2/choreo/connect/enforcer/util/TLSUtils.java
@@ -24,8 +24,8 @@ import io.grpc.netty.shaded.io.netty.handler.ssl.SslContext;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.wso2.choreo.connect.enforcer.commons.exception.EnforcerException;
 import org.wso2.choreo.connect.enforcer.config.ConfigHolder;
-import org.wso2.choreo.connect.enforcer.exception.EnforcerException;
 
 import java.io.BufferedInputStream;
 import java.io.ByteArrayInputStream;
