+++ enforcer-parent/enforcer/src/test/java/org/wso2/choreo/connect/enforcer/util/TLSUtilsTest.java
@@ -20,7 +20,7 @@ package org.wso2.choreo.connect.enforcer.util;
 
 import org.junit.Assert;
 import org.junit.Test;
-import org.wso2.choreo.connect.enforcer.exception.EnforcerException;
+import org.wso2.choreo.connect.enforcer.commons.exception.EnforcerException;
 
 import java.io.File;
 import java.io.IOException;
